feat: conditions rewrite WIP
