+++ src/lib/js/components/fields/edit-panel/edit-panel.js
@@ -1,9 +1,9 @@
 import i18n from '@draggable/i18n'
-import dom from '../../common/dom.js'
-import actions from '../../common/actions.js'
+import dom from '../../../common/dom.js'
+import actions from '../../../common/actions.js'
 import EditPanelItem from './edit-panel-item.mjs'
-import { capitalize } from '../../common/helpers.mjs'
-import { slugify, toTitleCase } from '../../common/utils/string.mjs'
+import { capitalize } from '../../../common/helpers.mjs'
+import { slugify, toTitleCase } from '../../../common/utils/string.mjs'
 
 /**
  * Element/Field class.
