+++ src/lib/js/components/fields/field.js
@@ -2,7 +2,7 @@ import i18n from '@draggable/i18n'
 import dom from '../../common/dom.js'
 import Panels from '../panels.js'
 import { clone, debounce, throttle, unique } from '../../common/utils/index.mjs'
-import EditPanel from './edit-panel.js'
+import EditPanel from './edit-panel/edit-panel.js'
 import Component from '../component.js'
 import { FIELD_CLASSNAME, CONDITION_TEMPLATE, ANIMATION_SPEED_BASE } from '../../constants.js'
 import Components from '../index.js'
